Separate logic into gotoraptor.ts, main entry point into index.ts
In tests, you don't want the main method to get called every time. Therefore the logic needs to be in a separate file so we can import that module in the tests without calling the main entry point.
